requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
kaggle>=1.5.16
matplotlib>=3.7.0
facenet-pytorch>=2.5.3
tqdm>=4.65.0
retinaface-pytorch>=0.0.8
scipy>=1.10.0
pytorch-lightning>=2.1.0
tensorboard>=2.15.0
transformers>=4.31.0
huggingface_hub>=0.16.4
efficientnet_pytorch>=0.7.1
